스킨 버전 가져오는 API
