+++ route/edit.py
@@ -92,13 +92,11 @@ def edit_2(conn, name):
                 data = get_data[0][0]
                 get_name = ''
 
-        js_data = edit_help_button()
-
         return easy_minify(flask.render_template(skin_check(), 
             imp = [name, wiki_set(), custom(), other2([' (' + load_lang('edit') + ')', 0])],
-            data =  get_name + js_data[0] + '''
+            data =  get_name + '''
                 <form method="post">
-                    ''' + js_data[1] + '''
+                    ''' + edit_button() + '''
                     <textarea id="content" rows="25" id="content" name="content">''' + html.escape(re.sub('\n$', '', data)) + '''</textarea>
                     <textarea style="display: none;" name="otent">''' + html.escape(re.sub('\n$', '', data_old)) + '''</textarea>
                     <hr class=\"main_hr\">
