+++ views/neo_yousoro/info.json
@@ -1,4 +1,5 @@
 {
+    "name" : "Neo_Yousoro",
     "skin_ver" : "v1.0.0",
     "require_ver" : "1"
 }
